Move TekuConfigurationBuilder into TekuConfiguration
